solved security vulnerability

+++ customer/customermethod.js
@@ -3,11 +3,10 @@ const config = require('../config');
 const misc = require('../extras/misc');
 
 const { con, sgst, cgst } = config;
-let menu = [];
+const menu = [];
 let resturantName; let
   gstin;
 let orderid = 0;
-
 exports.home = async (request, response) => {
   // eslint-disable-next-line max-len
 
@@ -27,33 +26,48 @@ exports.home = async (request, response) => {
       bill: request.session.order.orderbill,
     });
   } else {
-      menu.length = 0;
-    let sql = 'Select name, gst from client where fssai=? and active = 1';
-    const vars = [request.params.fssai];
-    let query = con.query({
-      sql,
+    const blockedSql = 'Select tableno, fssai from blocked where fssai=? and tableno = ?';
+    const blockedVars = [request.params.fssai, request.params.tableno];
+    const blockedQuery = con.query({
+      sql: blockedSql,
       timeout: 10000,
-    }, vars);
+    }, blockedVars);
+    let isBlocked = 0;
 
-    query.on('result', (result) => {
-    //    menu.push(result);
-      resturantName = result.name;
-      gstin = result.gst;
-      if (menu.length === 0) {
-        sql = `Select foodName, price, qty, category from menu_basic_${request.params.fssai} where acive=1 order by category`;
-        query = con.query({
-          sql,
-          timeout: 10000,
-        });
+    blockedQuery.on('result', () => { isBlocked = 1; response.send('blocked'); response.end(); });
+    if (isBlocked === 0) {
+      menu.length = 0;
+      let sql = 'Select name, gst from client where fssai=? and active = 1';
+      const vars = [request.params.fssai];
+      let query = con.query({
+        sql,
+        timeout: 10000,
+      }, vars);
 
-        query.on('result', (res) => {
-          menu.push(res);
-        });
-      }
-      response.render(path.join(`${__dirname}/customerhome.ejs`), { resturant: resturantName, tableno: request.params.tableno });
-    });
+      query.on('result', (result) => {
+      //    menu.push(result);
+        resturantName = result.name;
+        gstin = result.gst;
+        if (menu.length === 0) {
+          sql = `Select foodName, price, qty, category from menu_basic_${request.params.fssai} where acive=1 order by category`;
+          query = con.query({
+            sql,
+            timeout: 10000,
+          });
+
+          query.on('result', (res) => {
+            menu.push(res);
+          });
+        }
+        setTimeout(() => {
+          response.render(path.join(`${__dirname}/customerhome.ejs`), { resturant: resturantName, tableno: request.params.tableno });
+        }, 500);
+      });
+    }
+    //    });
   }
 };
+
 exports.begin = async (request, response) => {
   // eslint-disable-next-line max-len
   if (request.session.loggedin
@@ -71,28 +85,51 @@ exports.begin = async (request, response) => {
       bill: request.session.order.orderbill,
     });
   } else {
-    request.session.date = misc.today();
-    request.session.loggedin = true;
-    request.session.tableno = request.params.tableno;
-    request.session.fssai = request.params.fssai;
-    request.session.order = {
-      orderid: orderid + 1,
-      orderstatus: 'order',
-      orderdate: request.session.date = misc.today(),
-      ordercount: 0,
-      orderbill: 0,
-      orderdetails: {},
-    };
-    const sql2 = `delete from order_${request.params.fssai} where tableno = ?`;
-    const vars2 = [request.params.tableno];
-    const query2 = con.query({
-      sql: sql2,
+    const blockedSql = 'Select tableno, fssai from blocked where fssai=? and tableno = ?';
+    const blockedVars = [request.params.fssai, request.params.tableno];
+    const blockedQuery = con.query({
+      sql: blockedSql,
       timeout: 10000,
-    }, vars2);
-    query2.on('result', () => {});
-    response.render(path.join(`${__dirname}/customermenu.ejs`), {
-      resturant: resturantName, tableno: request.params.tableno, order: request.session.order, menu,
-    });
+    }, blockedVars);
+    let isblocked = 0;
+
+    blockedQuery.on('result', () => { isblocked = 1; response.send('blocked'); response.end(); });
+    if (isblocked === 0) {
+      request.session.date = misc.today();
+      request.session.loggedin = true;
+      request.session.tableno = request.params.tableno;
+      request.session.fssai = request.params.fssai;
+      request.session.order = {
+        orderid: orderid + 1,
+        orderstatus: 'order',
+        orderdate: request.session.date = misc.today(),
+        ordercount: 0,
+        orderbill: 0,
+        orderdetails: {},
+      };
+      const sql2 = `delete from order_${request.params.fssai} where tableno = ?`;
+      const vars2 = [request.params.tableno];
+      const query2 = con.query({
+        sql: sql2,
+        timeout: 10000,
+      }, vars2);
+      query2.on('result', () => {});
+
+      const blockSql = 'insert into blocked (fssai, tableno) values (?, ?) ';
+      const blockVars = [request.params.fssai, request.params.tableno];
+      const blockQuery = con.query({
+        sql: blockSql,
+        timeout: 10000,
+      }, blockVars);
+
+      blockQuery.on('result', () => {
+        // table blocked.
+      });
+
+      response.render(path.join(`${__dirname}/customermenu.ejs`), {
+        resturant: resturantName, tableno: request.params.tableno, order: request.session.order, menu,
+      });
+    }
   }
   orderid += 1;
   response.end();
@@ -116,11 +153,19 @@ exports.end = async (request, response) => {
     } else {
       request.session.destroy();
       menu.length = 0;
-      response.redirect(`/customer/${request.params.fssai}/${request.params.tableno}/begin`);
+      const blockedSQL = 'delete from blocked where fssai=? and tableno=?';
+      const blockedVars = [request.params.fssai, request.params.tableno];
+      const blockedQuery = con.query({
+        sql: blockedSQL,
+        timeout: 10000,
+      }, blockedVars);
+
+      blockedQuery.on('result', () => {});
+      response.redirect(`/customer/${request.params.fssai}/${request.params.tableno}/thankyou`);
       response.end();
     }
   } else {
-    response.redirect(`/customer/${request.params.fssai}/${request.params.tableno}/begin`);
+    response.redirect(`/customer/${request.params.fssai}/${request.params.tableno}/thankyou`);
     response.end();
   }
 };
